refactor: reorganizar servicios y separar responsabilidades
