+++ src/Components/Auth/register/register_hero.jsx
