+++ src/models/auth/login.model.js
@@ -1,0 +1,6 @@
+export const loginModel = (data) => {
+  return {
+    email: data.email,
+    password: data.password,
+  };
+};
