+++ src/pages/ForgotPassword.jsx
@@ -1,5 +1,5 @@
-import ForgotPasswordForm from "../Components/Auth/forgot_password_form";
-import ForgotPasswordHero from "../Components/Auth/forgot_password_hero";
+import ForgotPasswordForm from "../Components/Auth/forgot_password/forgot_password_form";
+import ForgotPasswordHero from "../Components/Auth/forgot_password/forgot_password_hero";
 import AuthLayout from "../layouts/AuthLayout";
 
 export default function ForgotPassword() {
