+++ src/pages/Login.jsx
@@ -1,5 +1,5 @@
-import LoginForm from "../Components/Auth/login_form";
-import LoginHero from "../Components/Auth/login_hero";
+import LoginForm from "../Components/Auth/login/login_form";
+import LoginHero from "../Components/Auth/login/login_hero";
 import AuthLayout from "../layouts/AuthLayout";
 
 export default function Login() {
