+++ src/pages/Register.jsx
@@ -1,5 +1,5 @@
-import RegisterForm from "../Components/Auth/register_form";
-import RegisterHero from "../Components/Auth/register_hero";
+import RegisterForm from "../Components/Auth/register/register_form";
+import RegisterHero from "../Components/Auth/register/register_hero";
 import AuthLayout from "../layouts/AuthLayout";
 
 export default function Register() {
