+++ src/services/auth/auth.service.js
@@ -1,4 +1,4 @@
-import API from "../api";
+import API from "../../api/api";
 
 export const registerUser = async (userData) => {
   try {
@@ -6,15 +6,6 @@ export const registerUser = async (userData) => {
     return response.data;
   } catch (error) {
     throw error.response?.data || { message: "Error en registro" };
-  }
-};
-
-export const getGeneros = async () => {
-  try {
-    const response = await API.get("/genero");
-    return response.data.data;
-  } catch (error) {
-    throw error.response?.data || { message: "Error al obtener los géneros" };
   }
 };
