+++ src/utils/jwt.utils.js
@@ -1,0 +1,8 @@
+export const decodeToken = (token) => {
+  try {
+    const payload = token.split(".")[1];
+    return JSON.parse(atob(payload));
+  } catch {
+    return null;
+  }
+};
