+++ src/utils/storage.js
@@ -1,0 +1,12 @@
+
+export const saveToken = (token) => {
+  localStorage.setItem("access_token", token);
+};
+
+export const getToken = () => {
+  return localStorage.getItem("access_token");
+};
+
+export const removeToken = () => {
+  localStorage.removeItem("access_token");
+};
